+++ consensus/hotstuff/basic/backend/backend.go
@@ -20,6 +20,7 @@ package backend
 
 import (
 	"crypto/ecdsa"
+	signer2 "github.com/ethereum/go-ethereum/consensus/hotstuff/signer"
 	"math/big"
 	"sync"
 	"time"
@@ -84,7 +85,7 @@ func New(config *hotstuff.Config, privateKey *ecdsa.PrivateKey, db ethdb.Databas
 	recentMessages, _ := lru.NewARC(inmemoryPeers)
 	knownMessages, _ := lru.NewARC(inmemoryMessages)
 
-	signer := NewSigner(privateKey, byte(hsc.MsgTypePrepareVote))
+	signer := signer2.NewSigner(privateKey, byte(hsc.MsgTypePrepareVote))
 	backend := &backend{
 		config: config,
 		//db:             db,
