+++ consensus/hotstuff/basic/backend/backend_test.go
@@ -1,7 +1,6 @@
 package backend
 
 import (
-	"bytes"
 	"math/big"
 	"testing"
 	"time"
@@ -9,63 +8,9 @@ import (
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/common/hexutil"
 	"github.com/ethereum/go-ethereum/consensus"
-	"github.com/ethereum/go-ethereum/consensus/hotstuff"
-	"github.com/ethereum/go-ethereum/consensus/hotstuff/basic/core"
 	"github.com/ethereum/go-ethereum/core/types"
-	"github.com/ethereum/go-ethereum/crypto"
 	"github.com/stretchr/testify/assert"
 )
-
-func TestSign(t *testing.T) {
-	s := newTestSigner()
-	data := []byte("Here is a string....")
-	sig, err := s.Sign(data)
-	assert.NoError(t, err, "error mismatch: have %v, want nil", err)
-
-	//Check signature recover
-	hashData := crypto.Keccak256(data)
-	pubkey, _ := crypto.Ecrecover(hashData, sig)
-	var signer common.Address
-	copy(signer[:], crypto.Keccak256(pubkey[1:])[12:])
-	assert.Equal(t, signer, getAddress(), "address mismatch: have %v, want %s", signer.Hex(), getAddress().Hex())
-}
-
-func TestCheckValidatorSignature(t *testing.T) {
-	vset, keys := newTestValidatorSet(5)
-
-	// 1. Positive test: sign with validator's key should succeed
-	data := []byte("dummy data")
-	hashData := crypto.Keccak256([]byte(data))
-	for i, k := range keys {
-		// Sign
-		sig, err := crypto.Sign(hashData, k)
-		assert.NoError(t, err, "error mismatch: have %v, want nil", err)
-
-		// CheckValidatorSignature should succeed
-		signer := NewSigner(k, 3)
-		addr, err := signer.CheckSignature(vset, data, sig)
-		assert.NoError(t, err, "error mismatch: have %v, want nil", err)
-
-		val := vset.GetByIndex(uint64(i))
-		assert.Equal(t, addr, val.Address(), "validator address mismatch: have %v, want %v", addr, val.Address())
-	}
-
-	// 2. Negative test: sign with any key other than validator's key should return error
-	key, err := crypto.GenerateKey()
-	assert.NoError(t, err, "error mismatch: have %v, want nil", err)
-
-	// Sign
-	sig, err := crypto.Sign(hashData, key)
-	assert.NoError(t, err, "error mismatch: have %v, want nil", err)
-
-	// CheckValidatorSignature should return ErrUnauthorizedAddress
-	signer := NewSigner(key, byte(core.MsgTypePrepareVote))
-	addr, err := signer.CheckSignature(vset, data, sig)
-	assert.Equal(t, err, hotstuff.ErrUnauthorizedAddress, "error mismatch: have %v, want %v", err, hotstuff.ErrUnauthorizedAddress)
-
-	emptyAddr := common.Address{}
-	assert.Equal(t, emptyAddr, common.Address{}, "address mismatch: have %v, want %v", addr, emptyAddr)
-}
 
 func TestPrepare(t *testing.T) {
 	chain, engine := singleNodeChain()
@@ -249,35 +194,4 @@ func TestPrepareExtra(t *testing.T) {
 	h.Extra = append(vanity, make([]byte, 15)...)
 	payload, err = emptySigner.PrepareExtra(h, valSet)
 	assert.Equal(t, expectedResult, payload)
-}
-
-func TestFillExtraAfterCommit(t *testing.T) {
-	vanity := bytes.Repeat([]byte{0x00}, types.HotstuffExtraVanity)
-	istRawData := hexutil.MustDecode("0xf858f8549444add0ec310f115a0e603b2d7db9f067778eaf8a94294fc7e8f22b3bcdcf955dd7ff3ba2ed833f8212946beaaed781d2d2ab6350f5c4566a2c6eaac407a6948be76812f765c24641ec63dc2852b378aba2b44080c0")
-	expectedCommittedSeal := append([]byte{1, 2, 3}, bytes.Repeat([]byte{0x00}, types.HotstuffExtraSeal-3)...)
-	expectedIstExtra := &types.HotstuffExtra{
-		Validators: []common.Address{
-			common.BytesToAddress(hexutil.MustDecode("0x44add0ec310f115a0e603b2d7db9f067778eaf8a")),
-			common.BytesToAddress(hexutil.MustDecode("0x294fc7e8f22b3bcdcf955dd7ff3ba2ed833f8212")),
-			common.BytesToAddress(hexutil.MustDecode("0x6beaaed781d2d2ab6350f5c4566a2c6eaac407a6")),
-			common.BytesToAddress(hexutil.MustDecode("0x8be76812f765c24641ec63dc2852b378aba2b440")),
-		},
-		Seal:          []byte{},
-		CommittedSeal: [][]byte{expectedCommittedSeal},
-	}
-	h := &types.Header{
-		Extra: append(vanity, istRawData...),
-	}
-
-	// normal case
-	assert.NoError(t, emptySigner.SealAfterCommit(h, [][]byte{expectedCommittedSeal}))
-
-	// verify istanbul extra-data
-	istExtra, err := types.ExtractHotstuffExtra(h)
-	assert.NoError(t, err)
-	assert.Equal(t, expectedIstExtra, istExtra)
-
-	// invalid seal
-	unexpectedCommittedSeal := append(expectedCommittedSeal, make([]byte, 1)...)
-	assert.Equal(t, errInvalidCommittedSeals, emptySigner.SealAfterCommit(h, [][]byte{unexpectedCommittedSeal}))
 }
