+++ consensus/hotstuff/signer/signer_test.go
@@ -1,0 +1,148 @@
+package signer
+
+import (
+	"bytes"
+	"crypto/ecdsa"
+	"sort"
+	"strings"
+	"testing"
+
+	"github.com/ethereum/go-ethereum/common"
+	"github.com/ethereum/go-ethereum/common/hexutil"
+	"github.com/ethereum/go-ethereum/consensus/hotstuff"
+	"github.com/ethereum/go-ethereum/consensus/hotstuff/basic/core"
+	"github.com/ethereum/go-ethereum/consensus/hotstuff/basic/validator"
+	"github.com/ethereum/go-ethereum/core/types"
+	"github.com/ethereum/go-ethereum/crypto"
+	"github.com/stretchr/testify/assert"
+)
+
+func TestSign(t *testing.T) {
+	s := newTestSigner()
+	data := []byte("Here is a string....")
+	sig, err := s.Sign(data)
+	assert.NoError(t, err, "error mismatch: have %v, want nil", err)
+
+	//Check signature recover
+	hashData := crypto.Keccak256(data)
+	pubkey, _ := crypto.Ecrecover(hashData, sig)
+	var signer common.Address
+	copy(signer[:], crypto.Keccak256(pubkey[1:])[12:])
+	assert.Equal(t, signer, getAddress(), "address mismatch: have %v, want %s", signer.Hex(), getAddress().Hex())
+}
+
+func TestCheckValidatorSignature(t *testing.T) {
+	vset, keys := newTestValidatorSet(5)
+
+	// 1. Positive test: sign with validator's key should succeed
+	data := []byte("dummy data")
+	hashData := crypto.Keccak256([]byte(data))
+	for i, k := range keys {
+		// Sign
+		sig, err := crypto.Sign(hashData, k)
+		assert.NoError(t, err, "error mismatch: have %v, want nil", err)
+
+		// CheckValidatorSignature should succeed
+		signer := NewSigner(k, 3)
+		addr, err := signer.CheckSignature(vset, data, sig)
+		assert.NoError(t, err, "error mismatch: have %v, want nil", err)
+
+		val := vset.GetByIndex(uint64(i))
+		assert.Equal(t, addr, val.Address(), "validator address mismatch: have %v, want %v", addr, val.Address())
+	}
+
+	// 2. Negative test: sign with any key other than validator's key should return error
+	key, err := crypto.GenerateKey()
+	assert.NoError(t, err, "error mismatch: have %v, want nil", err)
+
+	// Sign
+	sig, err := crypto.Sign(hashData, key)
+	assert.NoError(t, err, "error mismatch: have %v, want nil", err)
+
+	// CheckValidatorSignature should return ErrUnauthorizedAddress
+	signer := NewSigner(key, byte(core.MsgTypePrepareVote))
+	addr, err := signer.CheckSignature(vset, data, sig)
+	assert.Equal(t, err, hotstuff.ErrUnauthorizedAddress, "error mismatch: have %v, want %v", err, hotstuff.ErrUnauthorizedAddress)
+
+	emptyAddr := common.Address{}
+	assert.Equal(t, emptyAddr, common.Address{}, "address mismatch: have %v, want %v", addr, emptyAddr)
+}
+
+func TestFillExtraAfterCommit(t *testing.T) {
+	vanity := bytes.Repeat([]byte{0x00}, types.HotstuffExtraVanity)
+	istRawData := hexutil.MustDecode("0xf858f8549444add0ec310f115a0e603b2d7db9f067778eaf8a94294fc7e8f22b3bcdcf955dd7ff3ba2ed833f8212946beaaed781d2d2ab6350f5c4566a2c6eaac407a6948be76812f765c24641ec63dc2852b378aba2b44080c0")
+	expectedCommittedSeal := append([]byte{1, 2, 3}, bytes.Repeat([]byte{0x00}, types.HotstuffExtraSeal-3)...)
+	expectedIstExtra := &types.HotstuffExtra{
+		Validators: []common.Address{
+			common.BytesToAddress(hexutil.MustDecode("0x44add0ec310f115a0e603b2d7db9f067778eaf8a")),
+			common.BytesToAddress(hexutil.MustDecode("0x294fc7e8f22b3bcdcf955dd7ff3ba2ed833f8212")),
+			common.BytesToAddress(hexutil.MustDecode("0x6beaaed781d2d2ab6350f5c4566a2c6eaac407a6")),
+			common.BytesToAddress(hexutil.MustDecode("0x8be76812f765c24641ec63dc2852b378aba2b440")),
+		},
+		Seal:          []byte{},
+		CommittedSeal: [][]byte{expectedCommittedSeal},
+	}
+	h := &types.Header{
+		Extra: append(vanity, istRawData...),
+	}
+
+	// normal case
+	assert.NoError(t, emptySigner.SealAfterCommit(h, [][]byte{expectedCommittedSeal}))
+
+	// verify istanbul extra-data
+	istExtra, err := types.ExtractHotstuffExtra(h)
+	assert.NoError(t, err)
+	assert.Equal(t, expectedIstExtra, istExtra)
+
+	// invalid seal
+	unexpectedCommittedSeal := append(expectedCommittedSeal, make([]byte, 1)...)
+	assert.Equal(t, errInvalidCommittedSeals, emptySigner.SealAfterCommit(h, [][]byte{unexpectedCommittedSeal}))
+}
+
+var emptySigner = &SignerImpl{}
+
+type Keys []*ecdsa.PrivateKey
+
+func (slice Keys) Len() int {
+	return len(slice)
+}
+
+func (slice Keys) Less(i, j int) bool {
+	return strings.Compare(crypto.PubkeyToAddress(slice[i].PublicKey).String(), crypto.PubkeyToAddress(slice[j].PublicKey).String()) < 0
+}
+
+func (slice Keys) Swap(i, j int) {
+	slice[i], slice[j] = slice[j], slice[i]
+}
+
+func newTestValidatorSet(n int) (hotstuff.ValidatorSet, []*ecdsa.PrivateKey) {
+	// generate validators
+	keys := make(Keys, n)
+	addrs := make([]common.Address, n)
+	for i := 0; i < n; i++ {
+		privateKey, _ := crypto.GenerateKey()
+		keys[i] = privateKey
+		addrs[i] = crypto.PubkeyToAddress(privateKey.PublicKey)
+	}
+	vset := validator.NewSet(addrs, hotstuff.RoundRobin)
+	sort.Sort(keys) //Keys need to be sorted by its public key address
+	return vset, keys
+}
+
+func getAddress() common.Address {
+	return common.HexToAddress("0x70524d664ffe731100208a0154e556f9bb679ae6")
+}
+
+func getInvalidAddress() common.Address {
+	return common.HexToAddress("0x9535b2e7faaba5288511d89341d94a38063a349b")
+}
+
+func generatePrivateKey() (*ecdsa.PrivateKey, error) {
+	key := "bb047e5940b6d83354d9432db7c449ac8fca2248008aaa7271369880f9f11cc1"
+	return crypto.HexToECDSA(key)
+}
+
+func newTestSigner() hotstuff.Signer {
+	key, _ := generatePrivateKey()
+	return NewSigner(key, 3)
+}
